+++ lib/segment/src/segment/read_view/mod.rs
@@ -1,5 +1,6 @@
 mod deferred;
 mod facet;
+mod formula_rescore;
 mod order_by;
 mod payload;
 mod sampling;
